biometric

+++ src/viewmodels/useLoginViewModel.js
@@ -9,6 +9,7 @@ import {
   setRememberMe,
   setUserData,
   setWidgetToken,
+  setLoginCredentials,
 } from '../redux/authSlice';
 import {
   AuthorizationStatus,
@@ -156,13 +157,7 @@ const useLoginViewModel = () => {
       if (res?.data.UserName !== credentials?.userName) {
         dispatch(resetAllModules());
       }
-      dispatch(
-        setRememberMe({
-          userName: loginApiData?.userName,
-          password: loginApiData?.password,
-          rememberMe: checked,
-        }),
-      );
+
       if (res?.data?.policies?.length > 0) {
         dispatch(setPolicy(res?.data?.policies[0].policyNumber));
         dispatch(setSelectedPolicyObject(res?.data?.policies[0]));
@@ -173,6 +168,21 @@ const useLoginViewModel = () => {
           Data: res?.data,
         }),
       );
+      dispatch(
+        setLoginCredentials({
+          userName: loginApiData?.userName,
+          password: loginApiData?.password,
+        }),
+      );
+      if (rememberMe) {
+        dispatch(
+          setRememberMe({
+            userName: loginApiData?.userName,
+            password: loginApiData?.password,
+            rememberMe: true,
+          }),
+        );
+      }
     },
     onError: e => {
       dispatch(
@@ -219,37 +229,26 @@ const useLoginViewModel = () => {
   const handleLogin = async () => {
     const filled = LoginCheckForError();
     if (!filled) return;
+    if (isToggle && biometrics?.userName !== loginApiData?.userName) {
+      const rnBiometrics = new ReactNativeBiometrics();
+      await rnBiometrics.deleteKeys();
+
+      dispatch(SetIsToggle(false));
+      dispatch(setBiometrics(null));
+      dispatch(
+        setErrorModal({
+          Show: true,
+          message: 'Previous user biometric login is disabled.',
+        }),
+      );
+      return;
+    }
     let apiData = {
       userName: loginApiData?.userName,
       password: loginApiData?.password,
       deviceToken: deviceToken ?? '--',
       platform: 'mobileapp',
     };
-
-    if (isToggle) {
-      try {
-        const rnBiometrics = new ReactNativeBiometrics({
-          allowDeviceCredentials: true,
-        });
-        const {available, biometryType} =
-          await rnBiometrics.isSensorAvailable();
-        if (available) {
-          await rnBiometrics.deleteKeys();
-          await rnBiometrics.createKeys();
-          dispatch(
-            setBiometrics({
-              userName: loginApiData?.userName,
-              password: Buffer.from(loginApiData?.password, 'utf8').toString(
-                'base64',
-              ),
-              biometryType,
-            }),
-          );
-        }
-      } catch (error) {
-        console.log('Biometric setup error:', error);
-      }
-    }
     trigger(apiData);
   };
 
@@ -362,50 +361,19 @@ const useLoginViewModel = () => {
     navigation.navigate(screen, {step: 1, type: 'forgot'});
   };
 
-  const handleCheck = () => setChecked(!checked);
+  const handleCheck = () => {
+    const value = !checked;
+    setChecked(value);
+    dispatch(
+      setRememberMe({
+        rememberMe: value,
+        userName: loginApiData?.userName,
+        password: loginApiData?.password,
+      }),
+    );
+  };
 
   const tabs = ['login', 'signup'];
-
-  useEffect(() => {
-    if (
-      biometrics?.userName &&
-      loginApiData?.userName &&
-      biometrics?.userName !== loginApiData?.userName
-    ) {
-      dispatch(SetIsToggle(false));
-      dispatch(setBiometrics(null));
-      return;
-    }
-
-    if (isToggle && loginApiData?.userName && loginApiData?.password) {
-      const setupBiometrics = async () => {
-        try {
-          const rnBiometrics = new ReactNativeBiometrics({
-            allowDeviceCredentials: true,
-          });
-          const {available, biometryType} =
-            await rnBiometrics.isSensorAvailable();
-          if (available) {
-            await rnBiometrics.deleteKeys();
-            await rnBiometrics.createKeys();
-
-            dispatch(
-              setBiometrics({
-                userName: loginApiData?.userName,
-                password: Buffer.from(loginApiData?.password, 'utf8').toString(
-                  'base64',
-                ),
-                biometryType,
-              }),
-            );
-          }
-        } catch (error) {
-          console.log('Biometric setup error (toggle):', error);
-        }
-      };
-      setupBiometrics();
-    }
-  }, [isToggle, loginApiData?.userName, loginApiData?.password]);
 
   return {
     states: {
